#23 Moved AuthModelListItem component to the authentication directory.
